+++ src/swagger/client.ts
@@ -17,6 +17,7 @@ import {
 import type {
   GetFunctionalTestingExecutionTestParams,
   GetFunctionalTestingSuiteExecutionParams,
+  ListFunctionalTestingSuiteExecutionsParams,
   RunFunctionalTestingSuiteParams,
   RunFunctionalTestingTestParams,
 } from "./client/functional-testing-types";
@@ -372,6 +373,14 @@ export class SwaggerClient implements Client {
     args: GetFunctionalTestingExecutionTestParams,
   ): Promise<unknown> {
     return this.withFunctionalTesting((ftApi) => ftApi.getTestExecution(args));
+  }
+
+  async listFunctionalTestingSuiteExecutions(
+    args: ListFunctionalTestingSuiteExecutionsParams,
+  ): Promise<unknown> {
+    return this.withFunctionalTesting((ftApi) =>
+      ftApi.listSuiteExecutions(args),
+    );
   }
 
   async runFunctionalTestingSuite(